Improved Navbar, added CSS to App, moved the common folder into the components folder.
