+++ client/src/App.css
@@ -1,3 +1,5 @@
+@import 'colours';
+
 #root {
   max-width: 1280px;
   margin: 0 auto;
@@ -5,38 +7,23 @@
   text-align: center;
 }
 
-.logo {
-  height: 6em;
-  padding: 1.5em;
-  will-change: filter;
-  transition: filter 300ms;
-}
-.logo:hover {
-  filter: drop-shadow(0 0 2em #646cffaa);
-}
-.logo.react:hover {
-  filter: drop-shadow(0 0 2em #61dafbaa);
-}
-
-@keyframes logo-spin {
-  from {
-    transform: rotate(0deg);
-  }
-  to {
-    transform: rotate(360deg);
-  }
+nav {
+  background-color: var(--navbarBackground);
+  color: var(--navbarText);
+  display: flex;
+  justify-content: space-around;
+  align-items: start;
+  position: fixed;
+  top: 0px;
+  left: 0px;
+  padding: 1rem;
+  width: 100%;
 }
 
-@media (prefers-reduced-motion: no-preference) {
-  a:nth-of-type(2) .logo {
-    animation: logo-spin infinite 20s linear;
-  }
+nav a {
+  color: inherit;
 }
 
-.card {
-  padding: 2em;
-}
-
-.read-the-docs {
-  color: #888;
+main {
+  padding-top: 50px;  
 }
