+++ client/src/colours.css
@@ -1,0 +1,5 @@
+:root {
+    --navbarBackground: #000000;
+    --navbarText: #ffffff;
+    --background: #242424;
+  }
